requirements.txt
streamlit
sentence-transformers
pdfminer.six
scikit-learn
